+++ Example-OSX/Scenes/LayoutExampleSceneViewController+Configuration.swift
@@ -1,11 +1,3 @@
-//
-//  LayoutExampleSceneViewController+Configuration.swift
-//  Example-OSX
-//
-//  Created by Chris on 16/01/2019.
-//  Copyright © 2019 Christoffer Winterkvist. All rights reserved.
-//
-
 import Cocoa
 
 extension LayoutExampleSceneViewController {
@@ -57,6 +49,10 @@ extension LayoutExampleSceneViewController {
 }
 
 extension LayoutExampleSceneViewController {
+
+    func configureSceneLayoutTitle() {
+        currentBlueprintTitleTextField.stringValue = activeLayout.title
+    }
 
     func configureBluePrintLayout() {
         configureDynamicHeight()
